refactor: extract normalizeUrl to shared utility
